Feat: 게임 세팅 페이지 구현 진행중
현재 view만 구성되어진 상태
