+++ pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   cupertino_icons: ^1.0.2
   flip_card: ^0.6.0
   flip_panel: ^1.0.1
+  flutter_provider: ^2.1.0
+  provider: ^6.0.1
 
 dev_dependencies:
   flutter_test:
